requirements.txt
watchdog>=3.0.0
psutil>=5.9.0
customtkinter>=5.2.0
Pillow>=10.0.0
pyyaml>=6.0
